New tailored layouts for blue bar
